requirements.txt
alabaster==0.7.10
altgraph==0.16.1
captcha==0.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-allauth==0.41.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-formtools==2.2
django-import-export==2.0.2
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-pure-pagination==0.3.0
django-redis==4.11.0
django-redis-cache==2.1.0
django-reversion==3.0.7
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.9
jdcal==1.4.1
jsonfield==3.1.0
markdown2==2.3.8
MarkupPy==1.14
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.3
Pillow==7.0.0
psycopg2==2.8.5
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
swapper==1.1.2.post1
tablib==1.1.0
urllib3==1.25.8
xadmin2==2.0.3
xlrd==1.2.0
xlwt==1.3.0
